# backend/swagger.json
{"openapi":"3.0.0","paths":{"/users":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"Creates a new user and returns the created user data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}},"summary":"Create a new user","tags":["User"]},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":"Returns a list of all users","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/UserResponseDto"}]}}}}},"summary":"Get all users","tags":["User"]}},"/users/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns the user data for the given ID","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}},"summary":"Get a user by ID","tags":["User"]},"put":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"Updates the user data for the given ID and returns the updated user"}},"summary":"Update a user by ID","tags":["User"]},"delete":{"operationId":"UserController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Deletes the user with the given ID"}},"summary":"Delete a user by ID","tags":["User"]}}},"info":{"title":"Crud Caching API","description":"API documentation for the application","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","minLength":6},"name":{"type":"string"}},"required":["email","password","name"]},"UserResponseDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"email":{"type":"string","format":"email"},"createdAt":{},"updatedAt":{}},"required":["id","name","email","createdAt","updatedAt"]},"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string","minLength":3},"email":{"type":"string","format":"email"},"password":{"type":"string","minLength":6}}}}}}